service_1/requirements.txt
fastapi
uvicorn
python-multipart
kafka-python
alembic
psycopg2
joblib
scikit-learn